requirements.txt
# NVIDIA Canary-1B Flash Gradio Web UI Requirements

# Core ML and Audio Processing
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.21.0

# NVIDIA NeMo Framework and Dependencies
nemo_toolkit>=2.0.0
pytorch-lightning>=2.0.0
lightning>=2.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0
transformers>=4.0.0

# Web Interface
gradio>=4.0.0

# Additional Dependencies
tokenizers
safetensors
torchmetrics
aiohttp
regex
lhotse
einops
jiwer
editdistance
sentencepiece
sacremoses
Cython
webdataset
pyannote.audio
IPython
datasets
huggingface_hub

# Utilities
pathlib2
typing-extensions

# Optional: For better performance
accelerate

